Refactor project to clean architecture
