--- app/src/main/java/com/example/tibiaclone/data/local/FavoritesRepository.kt
@@ -1,4 +1,0 @@
-package com.example.tibiaclone.data.local
-
-class FavoritesRepository {
-}
