+++ app/src/main/java/com/example/tibiaclone/data/repository/FavoritesRepositoryImpl.kt
@@ -1,0 +1,7 @@
+package com.example.tibiaclone.data.repository
+
+import com.example.tibiaclone.domain.repository.FavoritesRepository
+
+class FavoritesRepositoryImpl : FavoritesRepository {
+    // TODO: implement favorites persistence using DataStore or Room
+}
