+++ app/src/main/java/com/example/tibiaclone/di/NetworkModule.kt
@@ -1,9 +1,12 @@
-package com.example.tibiaclone.data.network
+package com.example.tibiaclone.di
 
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import com.example.tibiaclone.data.remote.api.PokemonApi
+import com.example.tibiaclone.data.repository.PokemonRepositoryImpl
+import com.example.tibiaclone.domain.repository.PokemonRepository
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import javax.inject.Singleton
@@ -11,7 +14,7 @@ import javax.inject.Singleton
 @Module //This class will provide objects
 @InstallIn(SingletonComponent::class)
 
-object RetrofitInstance {
+object NetworkModule {
 
     @Provides
     @Singleton
@@ -24,6 +27,6 @@ object RetrofitInstance {
     @Provides
     @Singleton
     fun providePokemonRepository(pokemonApi: PokemonApi): PokemonRepository {
-        return PokemonRepository(pokemonApi)
+        return PokemonRepositoryImpl(pokemonApi)
     }
 }
