+++ app/src/main/java/com/example/tibiaclone/domain/model/PokemonType.kt
@@ -1,3 +1,5 @@
+package com.example.tibiaclone.domain.model
+
 import com.google.gson.annotations.SerializedName
 
 enum class PokemonType {
