+++ app/src/main/java/com/example/tibiaclone/domain/repository/FavoritesRepository.kt
@@ -1,0 +1,5 @@
+package com.example.tibiaclone.domain.repository
+
+interface FavoritesRepository {
+    // TODO: define methods for managing favorites
+}
