+++ app/src/main/java/com/example/tibiaclone/domain/repository/PokemonRepository.kt
@@ -1,0 +1,10 @@
+package com.example.tibiaclone.domain.repository
+
+import com.example.tibiaclone.domain.model.Pokemon
+
+interface PokemonRepository {
+    suspend fun getPokemon(id: Int): Pokemon
+    fun getPokemonFromCache(id: Int?): Pokemon?
+    suspend fun getListOfPokemons(quantity: Int): List<Pokemon>
+    suspend fun getFirst20Pokemons(): List<Pokemon>
+}
